Support for Servlet spec 2.3.
Remove all methods from SafeRequest and SafeResponse that do nothing
more then delegate to the wrapped request or response. All of these
methods are implemented in the parent HttpServletRequestWrapper and
HttpServletResponseWrapper respectively.

Add source and binary compatibility for
ServletResponse#setCharacterEncoding(String). This is done via reflection
in 2.4+ and just throws a UnsupportedOperationException in 2.3.

Extend TestHttpServletResponse to support tests relative to
setCharacterEncoding(String).

Add new test for SafeResponseTest in filters package to allow testing
of internal methods with package accessibility. Other tests for
SafeResponseTest have been left in reference.HTTPUtilsTest.