Relates to Issue 113. Tidied up full build process by introducing a dist profile for creating a distribution zip for the project (using the Maven Assembly plugin). The dist includes sources, libs, the site etc... To invoke you first need to increase the Maven heap size (if you haven't done so already e.g. "export MAVEN_OPTS=-Xmx512m") and then run "mvn clean site -Pdist". Once the build has completed you should find a zip file in the target named esapi-${version}-dist.zip.
Currently JAR signing is commented out like in the original pom. The main change I've made to this is to move the jks file to a top level resources directory and externalise the storepass etc... To get this working you need to remove the comments from the maven-jar-plugin and take the dist profile from resources/settings.xml and place it in your own Maven settings file. I did this so we could remove any sensitive information relating to the signing out of the pom.