Make Executor return ExecuteResult and fix deadlock in DefaultExecutor.
Executor previously returned a String of stdout only, which limited usefulness
for many use cases, including the unit tests which needed to be able to check
both stdout and stderr. Now both are returned to the client along with the
exit status because that is often useful, too.

The new method which has this behavior is named executeProgram(). The old
method executeSystemCommand() has been unchanged for backward compatibility.